intern_projects/Gen-AI-Travel-Chatbot/backend/requirements.txt
langchain>=0.1.17
langchain-community>=0.0.26
langchain-google-genai>=0.0.9
google-generativeai>=0.3.2
chromadb>=0.4.22
tiktoken>=0.5.1
python-dotenv>=1.0.1
PyPDF2>=3.0.1
geopy>=2.4.1
requests>=2.31.0
meteostat>=1.6.6
pydantic>=1.10.13
dateparser>=1.2.0
pypdf
